estudio_madame/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.14.0",
    "email-validator==2.2.0",
    "faker==33.1.0",
    "fastapi[standard]==0.115.6",
    "google-api-python-client==2.158.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "httpx==0.28.1",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.1",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.20",
    "sqlalchemy==2.0.36",
    "uvicorn[standard]==0.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
